default.json: bump frequency
Mend renovate does not run with enough frequency.
It looks like it runs once every 3 days on average.
Give a 3 days timeframe, so that at least one day is in the schedule interval.
We can bump this number by running our renovate bots, instead of relying on the share infrastructure.